feat: float mm values, free-position label editor, auto-RETRIEVE, OTA fix
- All printer mm settings (label, gap/Nachlauf, backfeed, offset) now use
  float to allow sub-mm precision (e.g. 9.8 mm)
- NVS storage migrated from UShort/char to float (getFloat/putFloat)
- API POST handler uses as<float>() for all mm fields
- renderLabelPreview() uses parseFloat for label/gap inputs
- Renamed "Abstand zwischen Etiketten" to "Nachlauf (Abreiß-Puffer)" to
  clarify it is post-label feed for tear-off, not gap between labels
- feedToNextLabel() simplified: totalMm = labelMm + nachlaufMm + offsetMm
- testPrint() prints a bordered frame matching the configured label height
- printLabel() reads 2D layout (x_mm, y_mm, size_mm, h_mm) from NVS
- Label editor redesigned: free 2D drag-and-drop with properties panel
- RETRIEVE state executes immediately (no confirmation), auto-dismisses 3s
- OTA: two-phase HTTPS download to handle GitHub→CDN redirect (HTTP -1 fix)
- Calibration tool: dead-zone (mm feed) + scale factor (mark-measure-save)

+++ src/ThermalPrinter.h
@@ -12,16 +12,16 @@ public:
     // Drucker-Konfiguration (NVS "printer")
     static uint32_t loadBaud();
     static void     saveBaud(uint32_t baud);
-    static uint16_t loadLabelMm();
-    static void     saveLabelMm(uint16_t mm);
-    static uint16_t loadGapMm();
-    static void     saveGapMm(uint16_t mm);
+    static float    loadLabelMm();
+    static void     saveLabelMm(float mm);
+    static float    loadGapMm();       // "Nachlauf" nach Etikett für Abreiß-Puffer
+    static void     saveGapMm(float mm);
     static bool     loadUseCut();
     static void     saveUseCut(bool cut);
-    static uint16_t loadBackfeedMm();
-    static void     saveBackfeedMm(uint16_t mm);
-    static int8_t   loadFeedOffsetMm();
-    static void     saveFeedOffsetMm(int8_t mm);
+    static float    loadBackfeedMm();
+    static void     saveBackfeedMm(float mm);
+    static float    loadFeedOffsetMm();
+    static void     saveFeedOffsetMm(float mm);
 
     // Kalibrierung (NVS "calib")
     static float    loadScaleFactor();
